requirements.txt
numpy==1.22.1
matplotlib==3.5.1
PyYAML==6.0